Add files via upload
In JSP (JavaServer Pages), creating a student login involves a few key steps. First, an HTML form is created to collect the student's username and password. This form submits data to a JSP file. In the JSP file, you retrieve the form data using request.getParameter(). Next, you connect to a database to validate the credentials using JDBC. If the credentials are valid, you establish a session and redirect the user to a welcome page; otherwise, you show an error message